feat(publish): upgrade to Dokka v2 and fix Maven Central publishing
- Upgrade Dokka from v1 to v2 (2.1.0)
- Replace dokkaJavadoc task with dokkaGenerate for v2 compatibility
- Fix NMCP plugin configuration for proper Maven Central publishing
- Add comprehensive development guide to README.md
- Remove obsolete PublishManExtension.kt file
- Update build configuration to use publishSonatypePublicationToCentralPortal

Key changes:
- Use :swc-binding:publishSonatypePublicationToCentralPortal for publishing
- Support USER_MANAGED publication type with manual approval
- Add detailed setup instructions for Maven Central Portal
- Fix repository configuration and credential handling

BREAKING CHANGE: Update publishing command from generic 'publish' to specific 'publishSonatypePublicationToCentralPortal'
