+++ build.gradle.kts
@@ -1,7 +1,16 @@
+import java.util.Properties
+
 plugins {
     alias(libs.plugins.jvm) apply false
     alias(libs.plugins.kotlin.serialization) apply false
     alias(libs.plugins.nmcp) apply false
+}
+
+// Load local.properties
+val localPropertiesFile = rootProject.file("local.properties")
+val localProperties = Properties()
+if (localPropertiesFile.exists()) {
+    localPropertiesFile.inputStream().use { localProperties.load(it) }
 }
 
 subprojects {
