+++ swc-binding/build.gradle.kts
@@ -1,12 +1,25 @@
+import java.util.Properties
+
 plugins {
     id("dev.yidafu.library")
     alias(libs.plugins.kotlin.serialization)
     id("org.jetbrains.kotlinx.kover") version "0.7.5"
+    alias(libs.plugins.dokka)
+
+    `maven-publish`
+    signing
+    alias(libs.plugins.nmcp)
 }
 
 group = "dev.yidafu.swc"
-version = "0.7.0"
+version = "0.8.0"
 
+// Load local.properties
+val localPropertiesFile = project.file("local.properties")
+val localProperties = Properties()
+if (localPropertiesFile.exists()) {
+    localPropertiesFile.inputStream().use { localProperties.load(it) }
+}
 kotlin {
     sourceSets {
         val main by getting {
@@ -33,11 +46,6 @@ tasks.test {
     useJUnitPlatform()
 }
 
-publishMan {
-    name.set("swc binding")
-    description.set("swc jvm binding by kotlin")
-}
-
 ktlint {
     filter {
         exclude { element ->
@@ -47,19 +55,103 @@ ktlint {
     }
 }
 
-// NMCP Plugin Configuration - credentials loaded by LibraryPlugin
-afterEvaluate {
-    nmcp {
-        publishAllPublications {
-            if (project.extensions.extraProperties.has("nmcp.username")) {
-                username.set(project.extensions.extraProperties.get("nmcp.username") as String?)
+val kotlinSourcesJar = project.tasks.findByName("kotlinSourcesJar") as org.gradle.jvm.tasks.Jar
+val dokkaJavadocJar = project.tasks.register<org.gradle.jvm.tasks.Jar>("dokkaJavadocJar") {
+    dependsOn("dokkaGenerate")
+    from(project.layout.buildDirectory.dir("dokka/html"))
+    archiveClassifier.set("javadoc")
+}
+// Maven Publishing 配置 - POM metadata for all publications
+// NMCP plugin automatically uses these publications for uploading to Central Portal
+publishing {
+    publications {
+        create<MavenPublication>("sonatype") {
+            // Create empty Javadoc JAR to satisfy Central Portal requirements
+            // Dokka HTML is available separately via dokkaHtml task
+      
+            from(project.components["java"])
+            artifact(kotlinSourcesJar)
+            artifact(dokkaJavadocJar)
+            versionMapping {
+                usage("java-api") {
+                    fromResolutionOf("runtimeClasspath")
+                }
+                usage("java-runtime") {
+                    fromResolutionResult()
+                }
             }
-            if (project.extensions.extraProperties.has("nmcp.password")) {
-                password.set(project.extensions.extraProperties.get("nmcp.password") as String?)
-            }
-            if (project.extensions.extraProperties.has("nmcp.publicationType")) {
-                publicationType.set(project.extensions.extraProperties.get("nmcp.publicationType") as String)
+
+            pom {
+                name.set("swc-binding")
+                description.set(
+                    "Swc Jvm Bdinding",
+                )
+                url.set("https://github.com/yidafu/swc-binding")
+
+                licenses {
+                    license {
+                        name.set("MIT License")
+                        url.set("https://opensource.org/licenses/MIT")
+                    }
+                }
+
+                developers {
+                    developer {
+                        id.set("yidafu")
+                        name.set("Yidafu")
+                        email.set("yidafu90@qq.com")
+                    }
+                }
+
+                scm {
+                    connection.set("scm:git:git://github.com/yidafu/swc-binding.git")
+                    developerConnection.set("scm:git:ssh://github.com/yidafu/swc-binding.git")
+                    url.set("https://github.com/yidafu/swc-binding")
+                }
             }
         }
     }
+}
+
+// NMCP Plugin 配置 - Modern plugin for Central Portal
+// See: https://github.com/GradleUp/nmcp
+nmcp {
+    // Publish to Central Portal
+    // Note: Central Portal does NOT support SNAPSHOT versions
+    // All versions must be release versions
+    publishAllPublications {
+        // Central Portal credentials (Portal user token)
+        username.set(
+            localProperties.getProperty("centralUsername")
+                ?: project.findProperty("centralUsername") as String?
+                ?: System.getenv("CENTRAL_USERNAME"),
+        )
+        password.set(
+            localProperties.getProperty("centralPassword")
+                ?: project.findProperty("centralPassword") as String?
+                ?: System.getenv("CENTRAL_PASSWORD"),
+        )
+
+        // Publication type: USER_MANAGED requires manual approval in Portal UI
+        // AUTOMATIC would auto-publish, but requires namespace verification
+        publicationType.set("USER_MANAGED")
+    }
+}
+
+// Signing configuration for Maven Central
+val signingKey = localProperties.getProperty("signing.key")
+    ?: System.getenv("SIGNING_KEY")
+val signingPassword = localProperties.getProperty("signing.password")
+    ?: System.getenv("SIGNING_PASSWORD")
+val signingKeyId = localProperties.getProperty("signing.keyId")
+    ?: System.getenv("SIGNING_KEY_ID")
+
+// Always enable signing for Maven Central publication
+signing {
+    // Use the GPG key from local properties
+    if (signingKey != null && signingPassword != null) {
+        useInMemoryPgpKeys(signingKey.trim(), signingPassword)
+    }
+    sign(publishing.publications)
+    isRequired = true
 }
